bookshelf/requirements.txt
aerospike==3.7.2
flask-pymongo==2.3.0
flask-sqlalchemy==2.4.0
flask==1.1.1
google-cloud-bigtable==0.33.0
google-cloud-datastore==1.8.0
google-cloud-storage==1.16.1
gunicorn==19.9.0
oauth2client==4.1.3
pymongo==3.8.0
pymysql==0.9.3
six>=1.13.0

